Make new palettes module
